Use rolepoint/logspout not gliderlabs/logspout.
It's not entirely clear which version of the logspout code the
gliderlabs repository is running.  In particular, the default logspout
image seems to be running an older version of the code that doesn't
support the IGNORE option (which we absolutely need to use).

To work around this, I've pushed up our own build of logspout at the
latest master version of the logspout github repository and pulled that
in.