Stop using a service to announce deis-syslog-ng.
Using a separate service to "announce" deis-syslog-ng was proving to be
a bit of a pain - you have to load the service in, and explicitly start
it along with deis-logger.

This commit moves the code for doing that into a shell script that can
run after deis-syslog-ng has started syslog-ng.